+++ src/lib/ngxf-uploader.module.ts
@@ -1,7 +1,7 @@
-import { NgModule, ModuleWithProviders } from '@angular/core';
+import { NgModule } from '@angular/core';
 
-import { NgxfSelectDirective } from './directive/ngxf-select.directive';
 import { NgxfDropDirective } from './directive/ngxf-drop.directive';
+import { NgxfSelectDirective } from './directive/ngxf-select.directive';
 
 @NgModule({
   declarations: [