+++ src/public_api.ts
@@ -1,6 +1,7 @@
 /*
  * Public API Surface of ngxf-uploader
  */
-export * from './lib/directive/ngxf-uploader.service';
+export * from './lib/ngxf-uploader.service';
 export * from './lib/ngxf-uploader.module';
+export * from './lib/ngxf-uploader.model';